requirements.txt
fastapi<0.69.0,>=0.68.0
uvicorn<0.16.0,>=0.15.0
openai>=1.0.0
python-dotenv>=0.19.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
alembic>=1.7.0
typer>=0.9.0 